backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic-settings==2.5.2
httpx==0.27.2
langfuse==2.36.2
pypdf==4.3.1
python-multipart==0.0.9
sqlite-vec==0.1.6
